Added wrapExpression as an extension point
Text interpolation is awesome, but it would be great to be able to do different things based on the contents - such as html interpolation. That particular case requires a modified html binding which supports virtual elements and uses a syntax that may not be universally wanted, so my commit only includes the extension point (which can be used to support a myriad of other use cases). A full example using a {{- html }} syntax for html interpolation can be found on my fiddle here: http://jsfiddle.net/daedalus28/DbEw7/